+++ lib/modules/import_data/import_skill_screenshot_page.dart
@@ -5,6 +5,7 @@ import 'package:chaldea/components/components.dart';
 import 'package:chaldea/modules/servant/servant_list_page.dart';
 import 'package:dio/dio.dart';
 import 'package:file_picker_cross/file_picker_cross.dart';
+import 'package:flutter/foundation.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:path/path.dart' as pathlib;
 
@@ -22,21 +23,37 @@ class ImportSkillScreenshotPageState extends State<ImportSkillScreenshotPage> {
   late Dio _dio;
   late List<File> imageFiles;
 
+  // update every build
+  Map<int, List<OneSvtRecResult>> resultsMap = {};
+
   @override
   void initState() {
     super.initState();
     imageFiles = db.runtimeData.svtRecognizeImageFiles;
-    _dio = Dio(BaseOptions(
-      baseUrl: 'http://localhost:8083',
-      // baseUrl: kServerRoot,
+    _dio = Dio(db.serverDio.options.copyWith(
+      // baseUrl: kDebugMode ? 'http://localhost:8083' : null,
       sendTimeout: 600 * 1000,
       receiveTimeout: 600 * 1000,
-      headers: HttpUtils.headersWithUA(),
+      headers: Map.from(db.serverDio.options.headers)
+        ..remove(Headers.contentTypeHeader),
     ));
   }
 
   @override
   Widget build(BuildContext context) {
+    resultsMap.clear();
+    results.forEach((e) {
+      if (e.svtNo != null) resultsMap.putIfAbsent(e.svtNo!, () => []).add(e);
+    });
+    int totalNum = results.length,
+        validNum = results.where((e) => e.isValid).length,
+        selectedNum = results.where((e) => e.isValid && e.checked).length,
+        dupNum = resultsMap.values
+            .where((e) =>
+                e.length > 1 &&
+                e.where((ee) => ee.isValid && ee.checked).length > 1)
+            .length;
+
     return Scaffold(
       appBar: AppBar(
         leading: BackButton(),
@@ -77,57 +94,55 @@ class ImportSkillScreenshotPageState extends State<ImportSkillScreenshotPage> {
                       ),
                     ),
                   ),
+                if (results.isNotEmpty)
+                  Padding(
+                    padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                    child: Text(
+                      LocalizedText.of(
+                          chs:
+                              '共$totalNum, 有效$validNum, 已选$selectedNum, 重复$dupNum',
+                          jpn:
+                              '合計$totalNum, 有効$validNum, 選択済み$selectedNum, 重複$dupNum',
+                          eng:
+                              'Total $totalNum, valid $validNum, selected $selectedNum, duplicated $dupNum'),
+                    ),
+                  ),
                 if (results.isNotEmpty) Expanded(flex: 2, child: _resultList()),
               ],
             ),
           ),
           kDefaultDivider,
-          _buildButtonBar(),
+          buttonBar,
         ],
       ),
     );
   }
 
-  Widget _buildButtonBar() {
+  Widget get buttonBar {
     List<OneSvtRecResult> usedResults =
         results.where((e) => e.isValid && e.checked).toList();
-    return Wrap(
-      crossAxisAlignment: WrapCrossAlignment.center,
-      alignment: WrapAlignment.center,
-      spacing: 6,
+    return ButtonBar(
+      alignment: MainAxisAlignment.center,
       children: [
-        ElevatedButton(
-            onPressed: imageFiles.isEmpty ? null : _uploadScreenshots,
-            child: Text(S.of(context).upload)),
-        ElevatedButton(
-            onPressed: _fetchResult,
-            child: Text(LocalizedText.of(
-                chs: '下载结果', jpn: '結果をダウンロード', eng: 'Download Result'))),
-        ElevatedButton(
-          child: Text(S.current.import_data),
-          onPressed: usedResults.isEmpty
-              ? null
-              : () => SimpleCancelOkDialog(
-                    title: Text(S.current.import_data),
-                    content: Text(
-                        '${usedResults.length}/${results.length} ${S.current.servant} -> '
-                        '${db.curUser.name} - ${S.current.plan_x(db.curUser.curSvtPlanNo + 1)}'),
-                    onTapOk: () {
-                      usedResults.forEach((result) {
-                        final status = db.curUser.svtStatusOf(result.svtNo!);
-                        status
-                          ..favorite = true
-                          ..curVal.skills = [
-                            result.skill1!,
-                            result.skill2!,
-                            result.skill3!
-                          ];
-                      });
-                      db.itemStat.update();
-                      EasyLoading.showSuccess(S.current.import_data_success);
-                    },
-                  ).showDialog(context),
-        ),
+        Wrap(
+          crossAxisAlignment: WrapCrossAlignment.center,
+          alignment: WrapAlignment.center,
+          spacing: 6,
+          children: [
+            ElevatedButton(
+                onPressed: imageFiles.isEmpty ? null : _uploadScreenshots,
+                child: Text(S.current.upload)),
+            ElevatedButton(
+                onPressed: _fetchResult,
+                child: Text(LocalizedText.of(
+                    chs: '下载结果', jpn: '結果をダウンロード', eng: 'Download Result'))),
+            ElevatedButton(
+              child: Text(S.current.import_data),
+              onPressed:
+                  usedResults.isEmpty ? null : () => _doImport(usedResults),
+            ),
+          ],
+        )
       ],
     );
   }
@@ -136,17 +151,23 @@ class ImportSkillScreenshotPageState extends State<ImportSkillScreenshotPage> {
     List<Widget> children = [];
     results.forEach((svtResult) {
       if (svtResult.imgBytes == null) return;
+
       Servant? svt;
       if (svtResult.svtNo != null) {
         svt = db.gameData.servantsWithUser[svtResult.svtNo];
       }
       bool valid = svt != null && svtResult.isValid;
+      int? dupNum = resultsMap[svtResult.svtNo]
+          ?.where((e) => e.isValid && e.checked)
+          .length;
       Widget nameBtn = TextButton(
         child: AutoSizeText(
           svt != null ? 'No.${svt.no} ${svt.info.localizedName}' : 'unknown',
           maxLines: 2,
           minFontSize: 6,
           maxFontSize: 14,
+          style: TextStyle(
+              color: dupNum != null && dupNum > 1 ? Colors.redAccent : null),
         ),
         style: TextButton.styleFrom(alignment: Alignment.centerLeft),
         onPressed: () async {
@@ -205,7 +226,15 @@ class ImportSkillScreenshotPageState extends State<ImportSkillScreenshotPage> {
           child: Row(
             mainAxisSize: MainAxisSize.min,
             children: [
-              Image.memory(svtResult.imgBytes!, width: 40),
+              GestureDetector(
+                child: Image.memory(svtResult.imgBytes!, width: 44),
+                onTap: () {
+                  SimpleCancelOkDialog(
+                    hideCancel: true,
+                    content: Image.memory(svtResult.imgBytes!),
+                  ).showDialog(context);
+                },
+              ),
               svt?.iconBuilder(context: context, width: 40) ??
                   db.getIconImage(null, width: 40, aspectRatio: 132 / 144),
             ],
@@ -284,8 +313,7 @@ class ImportSkillScreenshotPageState extends State<ImportSkillScreenshotPage> {
       String content = data2['msg'].toString();
       showInformDialog(context, title: title, content: content);
     } catch (e, s) {
-      print(e);
-      print(s);
+      logger.e('upload skill screenshots to server error', e, s);
       showInformDialog(context, title: 'Error', content: e.toString());
     } finally {
       EasyLoading.dismiss();
@@ -296,21 +324,47 @@ class ImportSkillScreenshotPageState extends State<ImportSkillScreenshotPage> {
     try {
       final response = await _dio.get('/downloadSvtResult',
           queryParameters: _getBaseAPIParams());
-      Map data = jsonDecode(response.data);
+      // Map data = jsonDecode(response.data);
       if (!mounted) return;
-      if (data['success'] == true) {
-        results = SvtRecResults.fromJson(Map.from(data['body'])).results;
+      final resp = ChaldeaResponse.fromResponse(response.data);
+
+      if (resp.success) {
+        results = SvtRecResults.fromJson(Map.from(resp.body)).results;
         // output = Map<String, int>.from(data['msg']);
         // output = Item.sortMapById(output);
         // print(output);
-        setState(() {});
+        if (results.isEmpty) {
+          EasyLoading.showInfo(LocalizedText.of(
+              chs: '识别结果为空',
+              jpn: '認識結果が空です',
+              eng: 'The recognition result is empty'));
+        }
+        if (mounted) setState(() {});
       } else {
         showInformDialog(context,
-            title: 'Response', content: data['msg'].toString());
+            title: 'Response', content: resp.msg.toString());
       }
     } catch (e) {
       showInformDialog(context, title: 'Error', content: e.toString());
     }
+  }
+
+  void _doImport(List<OneSvtRecResult> usedResults) {
+    SimpleCancelOkDialog(
+      title: Text(S.current.import_data),
+      content: Text('${usedResults.length} ${S.current.servant}'
+          ' -> ${db.curUser.name}'),
+      onTapOk: () {
+        usedResults.forEach((result) {
+          final status = db.curUser.svtStatusOf(result.svtNo!);
+          status
+            ..favorite = true
+            ..curVal.skills = [result.skill1!, result.skill2!, result.skill3!];
+        });
+        db.itemStat.update();
+        EasyLoading.showSuccess(S.current.import_data_success);
+      },
+    ).showDialog(context);
   }
 
   Widget get helpBtn {
@@ -320,25 +374,49 @@ class ImportSkillScreenshotPageState extends State<ImportSkillScreenshotPage> {
           chs: """0. 测试阶段：有任何问题欢迎反馈交流！！！
 1. 功能：解析截图中的从者技能等级（不包括灵基等级），应该不限服务器
 2. 使用方法: 
- - 点击右上角可同时导入多张截图
+ - 点击右上角可同时导入多张源截图
  - 上传成功后稍等片刻再下载结果
  - 每行识别结果从左到右分别是：
-   - 识别到的从者头像
+   - 识别到的从者头像: 点击可放大显示
    - 本地从者头像：可点击进入详情
    - 识别的从者编号姓名：点击进入列表页更改从者（若存在识别错误或识别失败）
-   - 三个技能数值：点击可修改数值
-   - 复选框：是否选择导入，仅识别出且技能数值正确时可用
+   - 三个技能数值：点击可分别修改数值
+   - 复选框：是否选择导入，仅识别出且技能数值正确时可勾选
+   - 如有重复出现的结果，将显示为红色，后者会覆盖前者!!!
+ - 点击导入按钮将识别结果导入到当前账户
 3. 必要条件!!!重要!!!
   - 技能升级页的截图，头像缩放最好为中或大，小头像识别失败率高。通过左下角切换。
   - 仅识别被“锁定锁定锁定”的从者，不会有人不锁吧
-  - 头像框务必完整显示，至少保证职阶图标、技能数字等完全显示
 4. 其他注意事项
  - 单次上传总大小有限制(~15MB)，否则会出现413错误，请分多次上传下载
  - 截图请勿裁剪等修改，请勿分屏
+ - 已经10/10/10并头像呈现灰色的大概率识别失败
  - 解析结果保留24h, 24h后可能删除
- - 新从者/灵衣需要服务器更新头像数据，如有未能识别的，请积极反馈""",
-          jpn: """""",
-          eng: """""",
+ - 新从者/灵衣数据可能更新不及时，如有未能识别的，请积极反馈""",
+          jpn: """""", //TODO
+          eng: """0. Test Stage: any question and feedback is welcomed!
+1. Feature: resolving servant skills from screenshots, adapted for all servers.
+2. Usage:
+  - click upper-right button to import multiple screenshots at same time
+  - "upload" to server then "download" results after a while
+  - every result row:
+    - cropped servant avatar: click to view the larger image
+    - identified servant avatar: click to view servant detail
+    - identified servant number and name: click to change servant if it's wrong or not identified
+    - identified skills: click to change skill level
+    - checkbox: whether to import, only enabled when servant and skills are valid
+    - duplicated servants will show their name in red font, the latter will override the former one when importing
+  - click "import" to merge results into current account
+3. Requirement/IMPORTANT
+  - screenshot of skill page: click scale button(left-bottom) in FGO to scale avatar to middle or large, the small scale may have a lower accuracy
+  - only *LOCKED* servants will be recognized
+4. Others
+  - The total size of a single upload is limited (~15MB), otherwise a 413 error will occur, please upload and download multiple times
+  - Don't crop/modify the screenshots
+  - recognition for servants whose skills are already 10/10/10 may be failed
+  - The result will be retained for about 24h, and may be deleted after 24h
+  - the data of latest servants or costumes may be missing, send feedback to me if so
+""",
         );
         SimpleCancelOkDialog(
           title: Text(S.of(context).help),